app/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy 
passlib
python-jose 
argon2-cffi
python-dotenv
alembic
python-jose[cryptography]
authlib
httpx
python-jose
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic[email]
python-multipart 
pyjwt